# docs/swagger.yaml
components:
  schemas:
...
    dto.ErrorResponse:
      properties:
        code:
          example: UNAUTHORIZED
          type: string
        error:
          example: unauthorized
          type: string
      type: object
...
    dto.Instance:
      properties:
        default_quality_profile_id:
          description: |-
            DefaultQualityProfileID / DefaultRootFolderPath — ADR-0009 S6. Carried on
            the list DTO so the Add-to-Sonarr modal can pre-fill defaults for the chosen
            instance without a second GET. omitempty: nil pointer = no default set.
          example: 1
          type: integer
        default_root_folder_path:
          example: /tv
          type: string
        health:
          enum:
          - Bootstrapping
          - Available
          - SelfThrottled
          - UnavailableAuth
          - UnavailableNetwork
          - UnavailableUnknown
          example: Available
          type: string
        last_check_at:
          type: string
        last_error:
          type: string
        mode:
          enum:
          - auto
          - manual
          example: auto
          type: string
        name:
          example: alpha
          type: string
        public_url:
          example: https://sonarr.example.com
          type: string
        transitions_count:
          type: integer
        url:
          example: http://sonarr:8989
          type: string
      type: object
    dto.InstanceCooldown:
      properties:
        guid_after_failed_grab_sec:
          type: integer
        guid_after_failed_import_sec:
          type: integer
        mode:
          enum:
          - smart
          - strict
          type: string
        series_after_grab_sec:
          type: integer
      type: object
...
    dto.InstanceCreateRequest:
      properties:
        api_key:
          example: abcd...
          type: string
        cooldown:
          $ref: '#/components/schemas/dto.InstanceCooldown'
        default_quality_profile_id:
          description: |-
            DefaultQualityProfileID / DefaultRootFolderPath — ADR-0009 S6 Add-to-Sonarr
            defaults. Optional pointers: omitted/null = clear/leave-unset. Soft hints —
            the application layer does NOT reject an unknown-but-non-negative profile id
            or a stale root-folder path (re-validated in the UI against live Sonarr).
          example: 1
          type: integer
        default_root_folder_path:
          example: /tv
          type: string
        dry_run:
          type: boolean
        health_check:
          $ref: '#/components/schemas/dto.InstanceHealthCheck'
        limits:
          $ref: '#/components/schemas/dto.InstanceLimits'
        mode:
          enum:
          - auto
          - manual
          example: auto
          type: string
        name:
          example: alpha
          maxLength: 64
          minLength: 1
          type: string
        parse_on_grab_enabled:
          description: |-
            ParseOnGrabEnabled — pointer so omitted is distinguishable from
            explicit false. Omitted/null defaults to true.
          example: true
          type: boolean
        public_url:
          description: |-
            PublicURL — optional browser-facing URL. Omitted/null = no
            override. Empty string is rejected at the application layer with
            INVALID_INSTANCE_PUBLIC_URL so a client cannot accidentally
            downgrade the override into a silent no-op.
          example: https://sonarr.example.com
          type: string
        ranking:
          $ref: '#/components/schemas/dto.InstanceRanking'
        rate_limit_burst:
          maximum: 10000
          minimum: 0
          type: integer
        rate_limit_rpm:
          maximum: 10000
          minimum: 0
          type: integer
        retry:
          $ref: '#/components/schemas/dto.InstanceRetry'
        scan_skip_handled_seasons:
          description: |-
            ScanSkipHandledSeasons — pointer so omitted is distinguishable
            from explicit false. Omitted/null defaults to true (matches the
            046b migration default). Concrete false disables the scan
            pre-filter for this instance, forcing every monitored season through
            the full evaluator.
          example: true
          type: boolean
        search:
          $ref: '#/components/schemas/dto.InstanceSearch'
        search_timeout_sec:
          maximum: 3600
          minimum: 0
          type: integer
        tags:
          $ref: '#/components/schemas/dto.InstanceTags'
        timeout_sec:
          maximum: 3600
          minimum: 0
          type: integer
        type:
          description: |-
            Type is the arr_instance.type discriminator. Optional on the wire —
            omitted/empty defaults to "sonarr" (backwards compatible). Ф6-R-6a.
          enum:
          - sonarr
          - radarr
          example: sonarr
          type: string
        url:
          example: http://sonarr:8989
          type: string
        webhook_install_enabled:
          description: |-
            WebhookInstallEnabled — pointer so omitted is distinguishable
            from explicit `false`. Omitted/null defaults to true (matches the
            migration default). Concrete false explicitly disables the 041c
            reconciler.
          example: true
          type: boolean
        webhook_url_override:
          description: |-
            WebhookURLOverride — optional base URL for the auto-installed
            Sonarr webhook. Omitted/null = use derived public URL from
            runtime config. Empty-string rejected as for PublicURL.
          example: https://seasonfill.example.com
          type: string
      required:
      - name
      - url
      type: object
    dto.InstanceDetail:
      properties:
        api_key:
          example: '***'
          type: string
        cooldown:
          $ref: '#/components/schemas/dto.InstanceCooldown'
        default_quality_profile_id:
          description: |-
            DefaultQualityProfileID / DefaultRootFolderPath — ADR-0009 S6 Add-to-Sonarr
            defaults. Always emitted; JSON `null` when no default is stored. Mirror
            PublicURL (no omitempty — a stable key so the SPA can branch on null).
          example: 1
          type: integer
        default_root_folder_path:
          example: /tv
          type: string
        dry_run:
          type: boolean
        health_check:
          $ref: '#/components/schemas/dto.InstanceHealthCheck'
        limits:
          $ref: '#/components/schemas/dto.InstanceLimits'
        mode:
          enum:
          - auto
          - manual
          example: auto
          type: string
        name:
          example: alpha
          type: string
        parse_on_grab_enabled:
          description: |-
            ParseOnGrabEnabled toggles the 044b parse-on-OnGrab hook. Always
            emitted as a concrete bool; defaults to true on fresh rows.
          example: true
          type: boolean
        public_url:
          description: |-
            PublicURL is the optional browser-facing URL (D64). Always
            emitted; the JSON value is `null` when no override is stored.
          example: https://sonarr.example.com
          type: string
        ranking:
          $ref: '#/components/schemas/dto.InstanceRanking'
        rate_limit_burst:
          type: integer
        rate_limit_rpm:
          type: integer
        retry:
          $ref: '#/components/schemas/dto.InstanceRetry'
        scan_skip_handled_seasons:
          description: |-
            ScanSkipHandledSeasons toggles the 046b scan pre-filter. Always
            emitted as a concrete bool; defaults to true on fresh rows. When
            true, scans skip seasons that Sonarr already handles (complete OR
            nothing-on-disk-and-aired). When false, every monitored season
            goes through the full evaluator (regression guard).
          example: true
          type: boolean
        search:
          $ref: '#/components/schemas/dto.InstanceSearch'
        search_timeout_sec:
          type: integer
        tags:
          $ref: '#/components/schemas/dto.InstanceTags'
        timeout_sec:
          type: integer
        type:
          enum:
          - sonarr
          - radarr
          example: sonarr
          type: string
        ui_url:
          description: |-
            UIURL is the derived browser-facing URL the SPA links to. Equals
            PublicURL when set, otherwise URL. Always emitted as a non-empty
            string so the SPA never has to compute the fallback itself.
          example: https://sonarr.example.com
          type: string
        updated_at:
          type: string
        url:
          example: http://sonarr:8989
          type: string
        webhook_install_enabled:
          description: |-
            WebhookInstallEnabled toggles the 041c reconciler (D65). Always
            emitted as a concrete bool; defaults to true on fresh rows.
          example: true
          type: boolean
        webhook_url_override:
          description: |-
            WebhookURLOverride is the optional base URL for the auto-installed
            Sonarr webhook (D65). Always emitted; `null` when unset.
          example: https://seasonfill.example.com
          type: string
      type: object
    dto.InstanceHealthCheck:
      properties:
        recheck_auth_sec:
          type: integer
        recheck_network_sec:
          type: integer
      type: object
    dto.InstanceLimits:
      properties:
        max_grabs_per_scan:
          type: integer
        scan_max_series:
          type: integer
      type: object
    dto.InstanceList:
      properties:
        instances:
          items:
            $ref: '#/components/schemas/dto.Instance'
          type: array
          uniqueItems: false
      type: object
...
    dto.InstanceRanking:
      properties:
        indexer_priority_enabled:
          type: boolean
        origin_bonus:
          type: number
      type: object
    dto.InstanceRetry:
      properties:
        initial_backoff_sec:
          type: integer
        max_attempts:
          type: integer
        max_backoff_sec:
          type: integer
      type: object
    dto.InstanceSearch:
      properties:
        min_custom_format_score:
          type: integer
        require_all_aired:
          type: boolean
        skip_anime:
          type: boolean
        skip_specials:
          type: boolean
      type: object
...
    dto.InstanceUpdateRequest:
      properties:
        api_key:
          example: abcd...
          type: string
        cooldown:
          $ref: '#/components/schemas/dto.InstanceCooldown'
        default_quality_profile_id:
          description: |-
            DefaultQualityProfileID / DefaultRootFolderPath — ADR-0009 S6 Add-to-Sonarr
            defaults. Optional pointers: omitted/null = clear/leave-unset. Soft hints —
            the application layer does NOT reject an unknown-but-non-negative profile id
            or a stale root-folder path (re-validated in the UI against live Sonarr).
          example: 1
          type: integer
        default_root_folder_path:
          example: /tv
          type: string
        dry_run:
          type: boolean
        health_check:
          $ref: '#/components/schemas/dto.InstanceHealthCheck'
        limits:
          $ref: '#/components/schemas/dto.InstanceLimits'
        mode:
          enum:
          - auto
          - manual
          example: auto
          type: string
        name:
          example: alpha
          maxLength: 64
          minLength: 1
          type: string
        parse_on_grab_enabled:
          description: |-
            ParseOnGrabEnabled — pointer so omitted is distinguishable from
            explicit false. Omitted/null defaults to true.
          example: true
          type: boolean
        public_url:
          description: |-
            PublicURL — optional browser-facing URL. Omitted/null = no
            override. Empty string is rejected at the application layer with
            INVALID_INSTANCE_PUBLIC_URL so a client cannot accidentally
            downgrade the override into a silent no-op.
          example: https://sonarr.example.com
          type: string
        ranking:
          $ref: '#/components/schemas/dto.InstanceRanking'
        rate_limit_burst:
          maximum: 10000
          minimum: 0
          type: integer
        rate_limit_rpm:
          maximum: 10000
          minimum: 0
          type: integer
        retry:
          $ref: '#/components/schemas/dto.InstanceRetry'
        scan_skip_handled_seasons:
          description: |-
            ScanSkipHandledSeasons — pointer so omitted is distinguishable
            from explicit false. Omitted/null defaults to true (matches the
            046b migration default). Concrete false disables the scan
            pre-filter for this instance, forcing every monitored season through
            the full evaluator.
          example: true
          type: boolean
        search:
          $ref: '#/components/schemas/dto.InstanceSearch'
        search_timeout_sec:
          maximum: 3600
          minimum: 0
          type: integer
        tags:
          $ref: '#/components/schemas/dto.InstanceTags'
        timeout_sec:
          maximum: 3600
          minimum: 0
          type: integer
        type:
          description: |-
            Type is the arr_instance.type discriminator. Optional on the wire —
            omitted/empty defaults to "sonarr" (backwards compatible). Ф6-R-6a.
          enum:
          - sonarr
          - radarr
          example: sonarr
          type: string
        url:
          example: http://sonarr:8989
          type: string
        webhook_install_enabled:
          description: |-
            WebhookInstallEnabled — pointer so omitted is distinguishable
            from explicit `false`. Omitted/null defaults to true (matches the
            migration default). Concrete false explicitly disables the 041c
            reconciler.
          example: true
          type: boolean
        webhook_url_override:
          description: |-
            WebhookURLOverride — optional base URL for the auto-installed
            Sonarr webhook. Omitted/null = use derived public URL from
            runtime config. Empty-string rejected as for PublicURL.
          example: https://seasonfill.example.com
          type: string
      required:
      - name
      - url
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: cookie
      name: seasonfill_session
      type: apiKey
...
info:
  description: |-
    REST API for seasonfill — Sonarr season-fill orchestration.
    Routes accept either X-Api-Key header or a signed
    seasonfill_session cookie issued by POST /api/v1/auth/login.
  title: Seasonfill API
  version: "1.0"
openapi: 3.1.0
paths:
  /admin/instances:
    get:
      description: Latest snapshot from the in-memory checker.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.InstanceList'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Unauthorized
      security:
      - CookieAuth: []
      - ApiKeyAuth: []
      summary: List Sonarr instance health
      tags:
      - instances
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.InstanceCreateRequest'
                description: Instance fields
                summary: body
        description: Instance fields
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.InstanceDetail'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Conflict
      security:
      - CookieAuth: []
      - ApiKeyAuth: []
      summary: Create a Sonarr instance
      tags:
      - instances
  /admin/instances/{name}:
    delete:
      parameters:
      - description: Instance name
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Not Found
      security:
      - CookieAuth: []
      - ApiKeyAuth: []
      summary: Delete a Sonarr instance and its history
      tags:
      - instances
    get:
      parameters:
      - description: Instance name
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.InstanceDetail'
          description: OK
          headers:
            Last-Modified:
              description: RFC1123 of updated_at
              schema:
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Not Found
      security:
      - CookieAuth: []
      - ApiKeyAuth: []
      summary: Get one Sonarr instance (masked)
      tags:
      - instances
    put:
      parameters:
      - description: Instance name
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.InstanceUpdateRequest'
                description: Instance fields
                summary: body
        description: Instance fields
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.InstanceDetail'
          description: OK
          headers:
            Last-Modified:
              description: RFC1123 of updated_at
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Not Found
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Precondition Failed
      security:
      - CookieAuth: []
      - ApiKeyAuth: []
      summary: Update a Sonarr instance
      tags:
      - instances
...
servers:
- description: Primary API base path
  url: /api/v1